Started on the UI Design
Started on the UI Design. Again probably should've done smaller commits,
but you live and you learn. Got the search page working, with a rough
search algorithm that should probably be replaced with a fuzzy finder,
however it does return the result of the page using an Enum. I haven't
modelled a programs behaviour like this before, but I'm starting to
really like rust's enums.