Updated the view UI
Username and Email are shown regardless of whether one is attached to
the account or not. Considering changing the Account struct to reflect
this. I've made this change so that the viewing UI can also be used for
editing, as I'll make the current line a textfield, meaning the user can
easily add a username or email if they wish, but it wouldn't work if
the fields were not there in the first place.