Users can now yank the current field into their clipboard
By pressing the "y" key in the view menu users can now copy the current
field into their clipboard. Strangely enough this only lasts until the
function returns, so currently it only stays in the clipboard until the
user presses a key inside of the program. This has the handy benefit of
being a security feature?