Fixed account data changes not being saved
Because I had a check for accounts sharing a label with another account
in the vec, changes would only ever be saved if the label of the account
was also changed. This was because the vec still contained the original
account that was being overwritten, meaning that it would always return
true if the label was not changed during editing.